% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_duplicates.R
\name{aggregate_duplicates}
\alias{aggregate_duplicates}
\title{Aggregation of duplicated columns in a data.frame}
\usage{
aggregate_duplicates(x, idcols = NULL, sumcols = NULL)
}
\arguments{
\item{x}{data.frame with}

\item{idcols}{names of columns identifying the locations which should be checked for duplication. If \code{NULL} all other columns than \code{sumcols} are used. Can't be \code{NULL} if \code{sumcols} is.}

\item{sumcols}{names of columns to be aggregated for identical ids. If \code{NULL} all other columns than \code{idcols} are used. Can't be \code{NULL} if \code{idcols} is.}
}
\value{
data.frame with columns as specified by \code{idcols} and \code{sumcols}.
Columns not (implicitly) given in these are dropped.
}
\description{
Aggregation of duplicated columns in a data.frame
}
\details{
\strong{WARNING}: Factors are replaced by strings in the result.
}
\examples{
# Example with a single duplicated location
x <- data.frame(municipality = c("A", "B", "B"), x = c(1, 2, 2), y = c(1, 1, 1),
                households = c(1, 1, 2), individuals = c(4, 3, 8))
# Three ways to achieve same aggregated version:
y1 <- aggregate_duplicates(x, idcols = c("municipality", "x", "y"),
                           sumcols = c("households", "individuals"))
y2 <- aggregate_duplicates(x, sumcols = c("households", "individuals"))
y3 <- aggregate_duplicates(x, idcols = c("municipality", "x", "y"))
y1
identical(y1, y2)
identical(y2, y3)
# Additional columns are dropped if not (possibly implicitly) referred to:
xx <- cbind(x, junk = 1:3)
xx
aggregate_duplicates(xx, idcols = c("municipality", "x", "y"),
                     sumcols = c("households", "individuals"))
}
