% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrtess.R
\name{aggrtess}
\alias{aggrtess}
\title{Initialize aggregation tessellation}
\usage{
aggrtess(X, centers, target, ...)
}
\arguments{
\item{X}{Point pattern with n_p points and n_m columns of marks}

\item{centers}{Point pattern of n_c tile centers or a string (see details).}

\item{target}{vector of length n_m}

\item{...}{passed to \code{make_aggrtess_centers} when \code{target} is a string..}
}
\value{
An aggregation tessellation of class \code{aggr_tess} which is a list with
the main components being:
\itemize{
\item X: Original marked point pattern
\item centers: Point pattern of n_c tile centers
\item tessel: tessellation with n_c tiles augmented with neighbour info
\item counts: n_m by n_c matrix with aggregates of mark values in each tile
Note: The aggregation tessellation does \strong{not} necessarily satisfy the aggregation
target at this point.
}
}
\description{
Initialize aggregation tessellation
}
\examples{
x <- aggrtess(frb, centers = "rectgridsample", target = c(50, 100))
x
}
