% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{aggrtess2sf}
\alias{aggrtess2sf}
\title{Conversion of aggrtess to sf format}
\usage{
aggrtess2sf(x, crs = 25832)
}
\arguments{
\item{x}{tessellation}

\item{crs}{crs to use with sf}
}
\value{
Simple features object with the polygons and their counts.
}
\description{
Conversion of aggrtess to sf format
}
\examples{

set.seed(42)
x <- aggrtess(frb, centers = "index_shift", target = c(50, 100))
x_sf <- aggrtess2sf(x)
if(interactive()){
  mapview::mapview(x_sf, zcol = c("households", "individuals"))
}
}
