% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented_dirichlet.R
\name{augmented_dirichlet}
\alias{augmented_dirichlet}
\title{Dirichlet tessellation augmented with neighbour info}
\usage{
augmented_dirichlet(X)
}
\arguments{
\item{X}{point pattern}
}
\value{
Dirichlet tessellation of class \code{tess} with additional field \code{neigh}
}
\description{
Dirichlet tessellation augmented with neighbour info
}
\examples{
set.seed(42)
frb_small <- rthin(frb, 0.01)
d <- augmented_dirichlet(frb_small)
head(d$neigh)
}
