% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrtess.R
\name{delete_tile}
\alias{delete_tile}
\title{Delete tile in aggregation tessellation}
\usage{
delete_tile(x, id)
}
\arguments{
\item{x}{Aggregation tessellation of class \code{aggrtess}.}

\item{id}{Integer indicating which tile to delete}
}
\value{
Aggregation tessellation. Either with one less tile than the input
(if split was successful) or with the same number of tiles (if split didn't succeed in \code{max_tries} attempts).
}
\description{
Delete tile in aggregation tessellation
}
\examples{
x <- aggrtess(frb, centers = "hexgridcenter", target = c(50, 100))
y <- delete_tile(x, 4)
}
