% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frb}
\alias{frb}
\title{Locations of households in Frederiksberg, Denmark}
\format{An object of class \code{ppp} of length 6.}
\usage{
frb
}
\description{
A dataset containing the coordinates of addresses in the Danish kommune
Frederiksberg as a marked planar point pattern (\code{ppp}). The marks are
artificial values of number of households and individuals at the location.
}
\keyword{datasets}
