% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrtess.R
\name{make_aggrtess_centers}
\alias{make_aggrtess_centers}
\title{Title}
\usage{
make_aggrtess_centers(X, centers, target, ..., adjust = 1)
}
\arguments{
\item{X}{Point pattern with n_p points and n_m columns of marks}

\item{centers}{String (see details).}

\item{target}{vector of length n_m}

\item{...}{ignored.}

\item{adjust}{adjustment of target. One or n_m numbers between 0 and 1.}
}
\value{
A point pattern of class \code{ppp}.
}
\description{
Title
}
\examples{
make_aggrtess_centers(frb, "rthin", c(50, 100))
}
