% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrtess.R
\name{markquadratcount}
\alias{markquadratcount}
\title{Sum of point pattern marks over tessellation}
\usage{
markquadratcount(X, tess)
}
\arguments{
\item{X}{point pattern of class \code{ppp}.}

\item{tess}{tessellation of class \code{tess}.}
}
\value{
Matrix of counts with one column for each tile in tessellation and
one row for each mark variable.
}
\description{
Sum of point pattern marks over tessellation
}
\examples{
set.seed(42)
frb_centers <- rthin(frb, 0.01)
frb_tess <- augmented_dirichlet(frb_centers)
counts <- markquadratcount(frb, frb_tess)
marks(frb_tess) <- t(counts)
if(interactive()){
  plot(frb_tess, do.col = TRUE)
}
}
