% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrtess.R
\name{plot.aggrtess}
\alias{plot.aggrtess}
\title{Plot aggregation tessellation of class aggrtess}
\usage{
\method{plot}{aggrtess}(
  x,
  ...,
  what = "counts",
  collapse = NULL,
  relative = FALSE,
  do.col = TRUE,
  do.labels = TRUE,
  main = ""
)
}
\arguments{
\item{x}{aggregation tessellation object of class \code{aggrtess}}

\item{...}{passed to \code{plot.tess} in spatstat.}

\item{what}{string to determine what to plot. Either \code{"counts"},
\code{"deviation"} or \code{"excess"}.}

\item{collapse}{function to use to collapse multiple values to a single value
if desired. Typically \code{sum} or \code{mean}.}

\item{relative}{logical to use relative values. Divides by number of tiles in
cluster for \code{excess} and divides by \code{x$target} otherwise.}

\item{do.col}{color each tile according to associated value. Defaults to \code{TRUE}.}

\item{do.labels}{annotate plot with tile labels (ids). Defaults to \code{TRUE}.}

\item{main}{main title of plot. Defaults to empty string.}
}
\value{
the marked tessellation of class \code{tess} which is plotted (invisibly).
}
\description{
Plot aggregation tessellation of class aggrtess
}
\examples{
if(interactive()){
x <- aggrtess(frb, centers = "rectgridsample", target = c(50, 100))
plot(x)
}
}
