% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrtess.R
\name{prune_aggrtess}
\alias{prune_aggrtess}
\title{Prune aggregation tessellation to remove all tiles below the target}
\usage{
prune_aggrtess(x, max_tiles = 100, ..., verbose = FALSE)
}
\arguments{
\item{x}{Aggregation tessellation of class \code{aggrtess}.}

\item{max_tiles}{Maximal number of tiles to delete (to ensure the computation ends in a resonable time if started in error on a very bad/big example.)}

\item{...}{Ignored.}

\item{verbose}{Logical to print diagnostics.}
}
\value{
Aggregation tessellation.
}
\description{
Prune aggregation tessellation to remove all tiles below the target
}
\examples{
X <- frb[seq(1, npoints(frb), by = 10)]
x <- aggrtess(X, centers = "hexgridcenter", target = c(50, 100))
y <- prune_aggrtess(x)
}
