% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrtess.R
\name{split_tile}
\alias{split_tile}
\title{Split tile in aggregation tessellation}
\usage{
split_tile(x, id, ..., max_tries = 100, verbose = FALSE)
}
\arguments{
\item{x}{Aggregation tessellation of class \code{aggrtess}.}

\item{id}{Integer id of the tile to try to split}

\item{...}{Ignored}

\item{max_tries}{Maximal number of times to try to split tile before giving up.}

\item{verbose}{Logical to print diagnostics.}
}
\value{
Aggregation tessellation. Either with one more tile than the input
(if split was successful) or with the same number of tiles (if split didn't succeed in \code{max_tries} attempts).
}
\description{
Split tile in aggregation tessellation
}
\examples{
X <- frb[seq(1, npoints(frb), by = 10)]
x <- aggrtess(X, centers = "hexgridcenter", target = c(50, 100))
x <- prune_aggrtess(x)
set.seed(42) # Reproducibility
y <- split_tile(x, 6)
}
