% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrtess.R
\name{target_deviation}
\alias{target_deviation}
\title{Find deviation from target for aggegation tessellation}
\usage{
target_deviation(x, ..., which = c("all", "bad", "worst"), relative = TRUE)
}
\arguments{
\item{x}{object of class \code{aggrtess}}

\item{...}{ignored}

\item{which}{character describing the desired tiles: \code{"all"} (all \code{n} tiles),
\code{"bad"} (0-\code{n} tiles not satisfying target), or \code{"worst"} (single tile).}

\item{relative}{logical to use relative deviations from target.}
}
\value{
\code{data.frame} with columns \code{id}, \code{min} and one additional column for
each mark. Number of rows is between 0 and \code{n} depending on \code{which}.
}
\description{
Find deviation from target for aggegation tessellation
}
\examples{
x <- aggrtess(frb, centers = "rectgridsample", target = c(50, 100))
target_deviation(x)
}
