% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrtess.R
\name{aggrtess}
\alias{aggrtess}
\title{Initialize aggregation tessellation}
\usage{
aggrtess(X, centers, target, ..., adjust = 1)
}
\arguments{
\item{X}{Point pattern with n_p points and n_m columns of marks}

\item{centers}{Point pattern of n_c tile centers or a string (see details).}

\item{target}{vector of length n_m}

\item{...}{ignored.}

\item{adjust}{adjustment of target. One or n_m numbers between 0 and 1.}
}
\value{
An aggregation tessellation of class \code{aggr_tess} which is a list with components:
\itemize{
\item X: Original marked point pattern
\item centers: Point pattern of n_c tile centers
\item cells: tessellation with n_c tiles augmented with neighbour info
\item counts: n_m by n_c matrix with aggregates of mark values in each tile
Note: The aggregation tessellation does \strong{not} necessarily satisfy the aggregation
target at this point.
}
}
\description{
Initialize aggregation tessellation
}
\examples{
aggrtess(frb, centers = "rthin", target = c(50, 100), )
}
